secureionic/Backend/requirements.txt
Django==3
psycopg2==2.8.4 --no-binary psycopg2
pillow==7.1.2
djangorestframework==3.11.0
django-filter==2.2.0
django-extensions==2.2.9
django-cors-middleware==1.5.0
django-utils2
six
djoser
djangorestframework_simplejwt
drf-yasg==1.20.0
django-sendgrid-v5